+++ packages/voice/__tests__/VoiceWebSocket.test.ts
@@ -7,13 +7,13 @@ beforeEach(() => {
 	WS.clean();
 });
 
-async function onceIgnoreError<T extends EventEmitter>(target: T, event: string) {
+async function onceIgnoreError<Emitter extends EventEmitter>(target: Emitter, event: string) {
 	return new Promise((resolve) => {
 		target.on(event, resolve);
 	});
 }
 
-async function onceOrThrow<T extends EventEmitter>(target: T, event: string, after: number) {
+async function onceOrThrow<Emitter extends EventEmitter>(target: Emitter, event: string, after: number) {
 	return new Promise((resolve, reject) => {
 		target.on(event, resolve);
 		setTimeout(() => reject(new Error('Time up')), after);